% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_score_heatmap.R
\name{confidence_score_heatmap}
\alias{confidence_score_heatmap}
\title{Confidence Score Heat Map}
\usage{
confidence_score_heatmap(confidence_scores = NULL, data_outcomes = NULL)
}
\arguments{
\item{confidence_scores}{A data frame of confidence scores generated from the aggregation methods in the form of \link{data_confidence_scores}.
Defaults to \link{data_confidence_scores} if no argument supplied.}

\item{data_outcomes}{A data frame of unique claims and the associated binary outcome in the form of \link{data_outcomes}. If no argument supplied
then defaults to \link{data_outcomes} supplied within package.}
}
\value{
Plot in viewer
}
\description{
Confidence scores displayed on a colour spectrum across generated methods and
assessed claims, split into predicted replication outcomes (TRUE/FALSE).
}
\examples{
\donttest{confidence_score_heatmap(data_confidence_scores, data_outcomes)}

}
