% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_score_ridgeplot.R
\name{confidence_score_ridgeplot}
\alias{confidence_score_ridgeplot}
\title{Confidence Score Ridge Plot}
\usage{
confidence_score_ridgeplot(confidence_scores = NULL)
}
\arguments{
\item{confidence_scores}{A data frame of confidence scores in long format in the form of \link{data_confidence_scores}}
}
\value{
A density ridge plot of aggregation methods
}
\description{
Display a ridge plot of confidence scores for each aggregation method
}
\examples{
\donttest{confidence_score_ridgeplot(data_confidence_scores)}

}
