% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_comment_ratings}
\alias{data_comment_ratings}
\title{Comment Votes for each commend id}
\format{
A tibble with 107 rows and 4 columns
\describe{
\item{comment_id}{character string of 91 unique IDs}
\item{vote}{numeric value, all '1'}
\item{user_name}{factor for anonymized IDs for each participant (43 participants included in this dataset)}
\item{timestamp}{character string for the date and time that the judgement was submitted, in the format of YYYY.MM.DD.H.M.S}
}
}
\usage{
data_comment_ratings
}
\description{
Comment Votes for each commend id
}
\keyword{datasets}
