% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_confidence_scores}
\alias{data_confidence_scores}
\title{Confidence Scores generated for 25 papers with 22 aggregation methods}
\format{
a tibble with 550 rows and 5 columns
\describe{
\item{method}{character string of method name}
\item{method_id}{integer of method IDs}
\item{paper_id}{character string of paper IDs}
\item{cs}{numeric of generated confidence scores}
\item{n_experts}{numeric of the number of expert judgements aggregated in confidence score}
}
}
\usage{
data_confidence_scores
}
\description{
Confidence Scores generated for 25 papers with 22 aggregation methods
}
\keyword{datasets}
