% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_justification_ratings}
\alias{data_justification_ratings}
\title{Justification ratings table with id}
\format{
A tibble with 1,967 rows and 4 columns
\describe{
\item{justification_id}{character string for specific justifications (1703 included here)}
\item{vote}{numeric column vote vote (either 1 or -1) }
\item{user_name}{factor for anonymized IDs for each participant (132 participants included in this dataset)}
\item{timestamp}{character string for the date and time that the judgement was submitted, in the format of YYYY.MM.DD.H.M.S}
}
}
\usage{
data_justification_ratings
}
\description{
Justification ratings table with id
}
\keyword{datasets}
