% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_outcomes}
\alias{data_outcomes}
\title{Replication outcomes for the papers}
\format{
a tibble with 25 rows and 2 columns
\describe{
\item{paper_id}{character string for the paper ID}
\item{outcome}{numeric value of replication outcome. 1 = replication success, 0 = replication failure}
}
}
\usage{
data_outcomes
}
\description{
Replication outcomes for the papers
}
\keyword{datasets}
