% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_supp_priors}
\alias{data_supp_priors}
\title{A table of prior means, to be fed into the BayPRIORsAgg aggregation method}
\format{
A tibble of 25 rows and 2 columns
\describe{
\item{paper_id}{character string with a unique id for each row corresponding to the assessed claim (from 125 papers total)}
\item{prior_means}{numeric with the average prior probability for the claim corresponding to the paper_id}
}
}
\usage{
data_supp_priors
}
\description{
A table of prior means, to be fed into the BayPRIORsAgg aggregation method
}
\keyword{datasets}
