% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_supp_QuizWAgg}
\alias{data_supp_QuizWAgg}
\title{A table of scores on the quiz to assess prior knowledge, to be fed into the QuizWAgg aggregation method}
\format{
A tibble 555 rows and 2 columns
\describe{
\item{user_name}{factor for anonymized IDs for each participant}
\item{quiz_score}{numeric for the participant's score on the quiz (min of 0, max of 16, NA if no questions answered)}
\item{quiz_score_even}{numeric score}
\item{quiz_score_stats}{Numeric stats score}
}
}
\usage{
data_supp_QuizWAgg
}
\description{
A table of scores on the quiz to assess prior knowledge, to be fed into the QuizWAgg aggregation method
}
\keyword{datasets}
