% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_users}
\alias{data_users}
\title{All user data from the platform}
\format{
a tibble with 550 rows and 3 columns
\describe{
\item{id}{character string with a unique ID for each row}
\item{username}{character string with the email address of the participant}
\item{displayName}{character string for the participant's avatar on the repliCATS platform}
}
}
\usage{
data_users
}
\description{
All user data from the platform
}
\keyword{datasets}
