% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_placeholder.R
\name{method_placeholder}
\alias{method_placeholder}
\title{Placeholder function with TA2 output}
\usage{
method_placeholder(expert_judgements, method_name)
}
\arguments{
\item{expert_judgements}{A data frame in the form of ratings}

\item{method_name}{Aggregation method to place into placeholder mode}
}
\value{
A tibble of confidence scores \code{cs} for each \code{paper_id}.
}
\description{
This functions stands in for when we haven't completed coding the method.
}
\details{
This function expects input from \link{preprocess_judgements} and outputs
for \link{postprocess_judgements}.
}
\examples{
\dontrun{method_placeholder(data_ratings, method_name = "TestMethod")}

}
