% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_judgements.R
\name{postprocess_judgements}
\alias{postprocess_judgements}
\title{Post-processing.}
\usage{
postprocess_judgements(method_output)
}
\arguments{
\item{method_output}{tibble created from one of the aggregation methods after
pre-processing' with columns for the aggregation \code{method}, \code{paper_id}, \code{aggregated_judgement} and
\code{n_experts}}
}
\value{
A tibble of confidence scores \code{cs} for each \code{paper_id}, corresponding
with an aggregation \code{method} (character).
}
\description{
Standardise the output from aggregation method's.
This function is called by every aggregation method as a final step.
}
