% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_reason.R
\name{weight_reason}
\alias{weight_reason}
\title{Weighting method: Total number of judgement reasons}
\usage{
weight_reason(expert_reasons)
}
\arguments{
\item{expert_reasons}{A dataframe in the form of \link{data_supp_reasons}}
}
\value{
A tibble of three columns \code{paper_id}, \code{user_name}, and \code{reason_count}
}
\description{
This function is used by \link{ReasoningWAgg} to weighting functions for the aggregation
type \code{"ReasonWAgg"}. Calculates weights based on the number of judgement reasoning
methods used by an individual
}
\details{
Individuals' weight is equal to the maximum number of judgement reasons given
}
\seealso{
Other weighting functions:
\code{\link[=weight_asym]{weight_asym()}},
\code{\link[=weight_interval]{weight_interval()}},
\code{\link[=weight_nIndivInterval]{weight_nIndivInterval()}},
\code{\link[=weight_outlier]{weight_outlier()}},
\code{\link[=weight_reason2]{weight_reason2()}},
\code{\link[=weight_varIndivInterval]{weight_varIndivInterval()}}
}
\concept{weighting functions}
