% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreCAT_package.R
\docType{package}
\name{aggreCAT}
\alias{aggreCAT-package}
\alias{aggreCAT}
\title{aggreCAT: mathematically aggregating expert judgements}
\description{
To learn more about \pkg{aggreCAT}, start with the vignettes: \code{vignette(package = "aggreCAT")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://replicats.research.unimelb.edu.au/}
  \item \url{https://github.com/metamelb-repliCATS/aggreCAT}
  \item \url{https://metamelb-replicats.github.io/aggreCAT/}
}

}
\author{
\strong{Maintainer}: David Wilkinson \email{david.wilkinson.research@gmail.com} (\href{https://orcid.org/0000-0002-9560-6499}{ORCID})

Authors:
\itemize{
  \item David Wilkinson \email{david.wilkinson.research@gmail.com} (\href{https://orcid.org/0000-0002-9560-6499}{ORCID})
  \item Elliot Gould (\href{https://orcid.org/0000-0002-6585-538X}{ORCID})
  \item Aaron Willcox (\href{https://orcid.org/0000-0003-2536-2596}{ORCID})
  \item Charles T. Gray
  \item Rose E. O'Dea (\href{https://orcid.org/0000-0001-8177-5075}{ORCID})
  \item Rebecca Groenewegen (\href{https://orcid.org/0000-0001-9177-8536}{ORCID})
}

}
\keyword{internal}
