% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreCAT_package.R
\docType{package}
\name{aggreCAT}
\alias{aggreCAT}
\alias{aggreCAT-package}
\title{aggreCAT: mathematically aggregating expert judgments}
\description{
To learn more about \pkg{aggreCAT}, start with the vignettes: \code{browseVignettes(package = "aggreCAT")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://replicats.research.unimelb.edu.au/}
}

}
\author{
\strong{Maintainer}: David Wilkinson \email{david.wilkinson.research@gmail.com} (\href{https://orcid.org/0000-0002-9560-6499}{ORCID})

Authors:
\itemize{
  \item Aaron Willcox \email{ aaron@willcox.io} (\href{https://orcid.org/0000-0003-2536-2596}{ORCID})
  \item Charles Gray (\href{https://orcid.org/00000-0002-9978-011X}{ORCID})
  \item Elliot Gould (\href{https://orcid.org/0000-0002-6585-538X}{ORCID})
  \item Anca Hanea (\href{https://orcid.org/0000-0003-3870-5949}{ORCID})
  \item Bonnie Wintle (\href{https://orcid.org/0000-0003-0236-6906}{ORCID})
  \item Rose E. O'Dea (\href{https://orcid.org/0000-0001-8177-5075}{ORCID})
}

}
\keyword{internal}
