% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_score_plots.R
\name{confidence_score_ridgeplot}
\alias{confidence_score_ridgeplot}
\title{Confidence Score Ridge Plot}
\usage{
confidence_score_ridgeplot(confidence_scores = NULL, size = 12, font = "serif")
}
\arguments{
\item{confidence_scores}{A data frame of confidence scores in long format in the form of \link{data_confidence_scores}}

\item{size}{Font size}

\item{font}{Set font family. Defaults to serif}
}
\value{
A density ridge plot of aggregation methods
}
\description{
Display a ridge plot of confidence scores for each aggregation method.
}
\examples{
\dontrun{confidence_scores_ridgeplot(data_confidence_scores, size = 12,
font = "sans",
soft_colour = "grey30")}


}
