% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_comments}
\alias{data_comments}
\title{data_comments}
\format{
A tibble with 2 rows and 11 columns
\describe{
\item{round}{character string, both 'round_1' (before discussion)}
\item{paper_id}{character string identifying 2 unique papers}
\item{user_name}{factor for anonymized IDs for two participants}
\item{question}{character string for the type of question, both 'comprehension'}
\item{justification_id}{character string identifying 2 unique justifications}
\item{comment_id}{character string identifying 2 unique comments}
\item{commenter}{redundant column, same as user_name}
\item{comment}{character string with free-text response for the user}
\item{vote_count}{numeric, both 0}
\item{vote_sum}{numeric, both 0}
\item{timestamp}{character string for the date and time that the judgement was submitted, in the format of YYYY.MM.DD.H.M.S}
}
}
\usage{
data_comments
}
\description{
data_comments
}
\keyword{datasets}
