% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_interval.R
\name{weight_interval}
\alias{weight_interval}
\title{Weighting method: Width of intervals}
\usage{
weight_interval(expert_judgements)
}
\arguments{
\item{expert_judgements}{A dataframe in the form of \link{data_ratings}}
}
\value{
A tibble in the form of the input \code{expert_judgements} argument with additional columns
supplying the calculated weight for each row's observation.
}
\description{
Calculates weights by interval width
}
\details{
This function is used inside \link{IntervalWAgg} for aggregation type \code{"IntWAgg"}. It
calculates the width of each three-point judgement (upper - lower), then returns
the weight as the inverse of this interval.
}
