% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_reason.R
\name{weight_reason}
\alias{weight_reason}
\title{Weighting method: Total number of judgement reasons}
\usage{
weight_reason(expert_reasons)
}
\arguments{
\item{expert_reasons}{A dataframe in the form of \link{data_supp_reasons}}
}
\description{
This function is used by \link{ReasoningWAgg} to calculate weights for the aggregation
type \code{"ReasonWAgg"}. Calculates weights based on the number of judgement reasoning
methods used by an individual
}
\details{
Individuals' weight is equal to the maximum number of judgement reasons given
}
